// src/sdk/generated/models/BulkOnboardResult.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BulkOnboardResult = {
  total: number;
  success: number;
  failed: number;
  errors: Array<Record<string, any>>;
  created_student_ids: Array<string>;
};